+++ tests/testthat/test-utils.R
@@ -1,0 +1,6 @@
+test_that("time_taken() works", {
+  expect_is(time_taken(), "proc_time") # first call is reset proc.time
+  expect_is(time_taken(), "character") # 2nd call is time taken
+  expect_is(time_taken(), "character") # 3rd call is time taken
+  expect_is(time_taken(TRUE), "proc_time") # param TRUE means reset
+})